+++ packages/admin-api-client/test/support/helpers.js
@@ -1,7 +1,10 @@
-import { configClient } from "../../index";
+const waitOn = require("wait-on");
+
+import { configClient } from "../../src/index";
 
 configClient({
   port: 3110,
+  host: "127.0.0.1",
 });
 
 const wait = (time = 1000) => {
@@ -12,6 +15,11 @@ const wait = (time = 1000) => {
   });
 };
 
+const waitForServer = (port) => {
+  return waitOn({ resources: [`tcp:127.0.0.1:${port}`] });
+};
+
 module.exports = {
   wait,
+  waitForServer,
 };