+++ packages/cypress-commands/src/index.ts
@@ -1,0 +1,8 @@
+export * from "./types";
+export { AdminApiClient } from "./AdminApiClient";
+
+import { register as registerer } from "./register";
+
+export function register() {
+  registerer(Cypress, cy);
+}