+++ packages/cypress-commands/test/Cypress.mock.js
@@ -1,6 +1,5 @@
-const sinon = require("sinon");
-
-const Mock = class Mock {
+import sinon from "sinon";
+export default class Cypress {
   constructor() {
     this._sandbox = sinon.createSandbox();
 
@@ -9,16 +8,25 @@ const Mock = class Mock {
         add: this._sandbox.stub(),
       },
       env: this._sandbox.stub(),
+      Promise: Promise,
+    };
+
+    this._cyStubs = {
+      wrap: this._sandbox.stub().callsFake((promise) => {
+        return promise;
+      }),
+      log: this._sandbox.stub(),
     };
   }
 
   get stubs() {
-    return this._stubs;
+    return {
+      Cypress: this._stubs,
+      cy: this._cyStubs,
+    };
   }
 
   restore() {
     this._sandbox.restore();
   }
-};
-
-module.exports = Mock;
+}