+++ packages/cypress-commands/test/register.spec.js
@@ -1,13 +1,13 @@
-const CypressMock = require("./Cypress.mock");
+import CypressMock from "./Cypress.mock";
 
-const register = require("../src/register");
+import { register } from "../src/register";
 
 describe("register", () => {
   let cypressMock;
 
   beforeEach(() => {
     cypressMock = new CypressMock();
-    register(cypressMock.stubs);
+    register(cypressMock.stubs.Cypress, cypressMock.stubs.cy);
   });
 
   afterEach(() => {
@@ -16,27 +16,31 @@ describe("register", () => {
 
   describe("commands", () => {
     it("should register mocksConfigAdminApiClient command", () => {
-      expect(cypressMock.stubs.Commands.add.calledWith("mocksConfigAdminApiClient")).toBe(true);
+      expect(cypressMock.stubs.Cypress.Commands.add.calledWith("mocksConfigAdminApiClient")).toBe(
+        true
+      );
     });
 
     it("should register mocksSetCollection command", () => {
-      expect(cypressMock.stubs.Commands.add.calledWith("mocksSetCollection")).toBe(true);
+      expect(cypressMock.stubs.Cypress.Commands.add.calledWith("mocksSetCollection")).toBe(true);
     });
 
     it("should register mocksSetDelay command", () => {
-      expect(cypressMock.stubs.Commands.add.calledWith("mocksSetDelay")).toBe(true);
+      expect(cypressMock.stubs.Cypress.Commands.add.calledWith("mocksSetDelay")).toBe(true);
     });
 
     it("should register mocksSetConfig command", () => {
-      expect(cypressMock.stubs.Commands.add.calledWith("mocksSetConfig")).toBe(true);
+      expect(cypressMock.stubs.Cypress.Commands.add.calledWith("mocksSetConfig")).toBe(true);
     });
 
     it("should register mocksUseRouteVariant command", () => {
-      expect(cypressMock.stubs.Commands.add.calledWith("mocksUseRouteVariant")).toBe(true);
+      expect(cypressMock.stubs.Cypress.Commands.add.calledWith("mocksUseRouteVariant")).toBe(true);
     });
 
     it("should register mocksRestoreRouteVariants command", () => {
-      expect(cypressMock.stubs.Commands.add.calledWith("mocksRestoreRouteVariants")).toBe(true);
+      expect(cypressMock.stubs.Cypress.Commands.add.calledWith("mocksRestoreRouteVariants")).toBe(
+        true
+      );
     });
   });
 });