+++ packages/cypress-commands/CHANGELOG.md
@@ -10,6 +10,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 ### Removed
 
+## [6.0.0] - 2022-08-04
+
+### Added
+- feat(#229): Support managing multiple Mocks Servers allowing to create multiple ApiClients
+- feat(#229): Export `MocksServerApiClient` class for creating api clients
+- feat(#230): Add Cypress logs when running commands and MOCKS_SERVER_LOGS environment variable for disabling them
+
+### Changed
+- feat: BREAKING CHANGE. Export `register` method for registering commands.
+- refactor: Migrate to TypeScript
+
 ## [5.0.0] - 2022-07-22
 
 ### Changed