+++ packages/main/CHANGELOG.md
@@ -14,6 +14,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 ### Breaking change
 
+## [3.8.0] - 2022-08-04
+
+### Changed
+- chore(deps): Update core dependency to v3.8.0
+
 ## [3.7.0] - 2022-07-29
 
 ### Changed